make it familiar with PEP 8 and put the validators in separated file
