+++ Carshering/settings.py
@@ -142,5 +142,3 @@ STATICFILES_DIRS = [
 # https://docs.djangoproject.com/en/4.1/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
-
-RENT_LENGTH_IN_DAYS = 30
