+++ carservice/forms.py
@@ -2,36 +2,23 @@ from . models import Car, Offer, Rent
 from django import forms
 
 
-
 class CarUpdateForm(forms.ModelForm):
     class Meta:
         model = Car
         fields = ['vin', 'car_mileage', 'car_brand', 'car_model', 'date_of_prod', "car_photo"]
 
 
-class CarDeleteForm(forms.Form):
-    car = forms.ChoiceField(label='Car')
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields['car'].choices = [(car.id, str(car)) for car in Car.objects.all()]
-
 class OfferUpdateForm(forms.ModelForm):
     class Meta:
         model = Offer
         fields = ['description', 'price']
 
-class OfferDeleteForm(forms.Form):
-    offer = forms.ChoiceField(label='Offer')
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields['offer'].choices = [(offer.id, str(offer)) for offer in Offer.objects.all()]
 
 class RentUpdateForm(forms.ModelForm):
     class Meta:
         model = Rent
         fields = ['status', 'rent_start', 'offer', 'user']
+
 
 class RentDeleteForm(forms.Form):
     def __init__(self, user, *args, **kwargs):
@@ -39,5 +26,3 @@ class RentDeleteForm(forms.Form):
         self.fields['rent'].queryset = Rent.objects.filter(user=user)
 
     rent = forms.ModelChoiceField(queryset=Rent.objects.none())
-
-
