+++ carservice/templates/car_read.html
@@ -13,7 +13,7 @@
 {% for car in cars %}
 <div class="car-item" style="display: flex ; margin-bottom: 24px; align-items: flex-end;">
     {% if car.car_photo %}
-        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 20px;" src="{{ car.car_photo.url }}">
+        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 20px;" src="{{ car.car_photo.url }}" alt="Car photo">
             <div>
                 <p style="margin-left: 20px;">{{ car.car_brand }} {{ car.car_model }}</p>
                 <p style="margin-left: 20px;"> Mileage: {{ car.car_mileage }} KM</p>
