+++ carservice/templates/offer_create.html
@@ -1,7 +1,10 @@
 {% extends 'base.html' %}
+{% load crispy_forms_filters %}
 {% load static %}
 {% load crispy_forms_tags %}
 {% block body %}
+
+
 <h3 style="margin-top: 10px;">Create offer:</h3>
     <hr>
 <form method="post">{% csrf_token %}
@@ -9,4 +12,4 @@
     <legend style="font-size: 12px; color: gray; margin-top: -11px;">*Required fields</legend>
     <input class="btn btn-secondary" type="submit" value="CREATE">
 </form>
-{% endblock body%}
+{% endblock body %}
