+++ carservice/templates/offer_detail.html
@@ -17,9 +17,11 @@
     <h2 style="margin-top: 10px;">OFFER DETAILS</h2>
     <hr>
     <div class="offer-details-container">
-        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ offer.car.car_photo.url }}">
+        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ offer.car.car_photo.url }}" alt="Car photo">
         <div style="flex-direction: column;">
-            <textarea class="form-control description-field" readonly>{{ offer.description }}</textarea>
+            <label>
+                <textarea class="form-control description-field" readonly>{{ offer.description }}</textarea>
+            </label>
             <p style="margin-bottom: 0;">Shared by {{ offer.user }}</p>
             <a href="{% url 'all_offers' %}" class="btn btn-outline-secondary" style="padding: 3px; margin-top: 5px;">PREVIOUS PAGE</a>
         </div>
