+++ carservice/templates/offer_detail_search.html
@@ -17,9 +17,9 @@
     <h2 style="margin-top: 10px;">OFFER DETAILS</h2>
     <hr>
     <div class="offer-details-container">
-        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ offer.car.car_photo.url }}">
+        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ offer.car.car_photo.url }}" alt="Car photo">
         <div style="flex-direction: column;">
-            <textarea class="form-control description-field" readonly>{{ offer.description }}</textarea>
+                <textarea class="form-control description-field" readonly>{{ offer.description }}</textarea>
             <a href="{{ request.META.HTTP_REFERER }}" class="btn btn-outline-secondary" style="padding: 3px; margin-top: 5px;">PREVIOUS PAGE</a>
         </div>
     </div>
