+++ carservice/templates/offer_result.html
@@ -4,7 +4,7 @@
 {% block body %}
     <h1>Offer Result</h1>
     <div style="display: flex; align-items: flex-end; background-color: #f8f8f8; padding: 15px; border-radius: 10px;">
-        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ car.car_photo.url }}">
+        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ car.car_photo.url }}" alt="Car photo">
         <div>
             <p style="margin-left: 2px;">{{ car.car_brand }} {{ car.car_model }} for {{ offer.price }} PLN/day.</p>
             <div style="display: flex; align-items: flex-end;">
