+++ carservice/templates/rent_create.html
@@ -1,4 +1,5 @@
 {% extends 'base.html' %}
+{% load crispy_forms_filters %}
 {% load static %}
 {% load crispy_forms_tags %}
 
@@ -23,7 +24,7 @@
     {% csrf_token %}
     {{ form.rent_start|as_crispy_field }}
     {{ form.duration|as_crispy_field }}
-      <legend style="font-size: 12px; color: gray; margin-top: -11px;">*Required fields</legend>
+      <legend style="font-size: 12px; color: gray; margin-top: -11px;">*Required fields (Maximum duration is 30 days)</legend>
     <input type="submit" class="btn btn-secondary" value="Rent">
     <a href="{% url 'offer_read' %}" class="btn btn-outline-secondary" style="margin: 0 20px 0 20px;">Back</a>
   </form>
