+++ carservice/templates/rent_detail.html
@@ -5,7 +5,7 @@
     <h2 style="margin-top: 10px;">RENT DETAILS</h2>
     <hr>
     <div style="display: flex; align-items: flex-end;">
-        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ rent.offer.car.car_photo.url }}">
+        <img class="rounded float-start" style="display: flex; max-width: 200px; margin-right: 10px;" src="{{ rent.offer.car.car_photo.url }}" alt="Car photo">
         <div style="flex-direction: column;">
             <p style="margin-bottom: 0;">Status: {{ rent.status }}</p>
             <p style="margin-bottom: 0;">Rent Start: {{ rent.rent_start }}</p>
