+++ users/models.py
@@ -8,6 +8,3 @@ class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     image = models.ImageField(upload_to='media/static/image', default='media/static/image/avatar.png')
     location = models.CharField(max_length=30, null=True, blank=True)
-
-
-
